# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bridgepath Network API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /cv-reviews:
    post:
      operationId: createCvReview
      tags: [cv-reviews]
      summary: Submit CV for AI review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCvReviewBody"
      responses:
        "201":
          description: CV review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvReview"
...
  /cv-reviews/{reviewId}:
    get:
      operationId: getCvReview
      tags: [cv-reviews]
      summary: Get a CV review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CV review found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvReview"
  /cv-reviews/{reviewId}/human-review:
    post:
      operationId: requestHumanReview
      tags: [cv-reviews]
      summary: Request human HR review (paid upgrade)
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestHumanReviewBody"
      responses:
        "200":
          description: Human review requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvReview"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [job_seeker, employer, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        experience:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        portfolioUrl:
          type: string
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        companyWebsite:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - skills
        - createdAt
        - updatedAt
...
    Job:
      type: object
      properties:
        id:
          type: integer
        employerId:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
          nullable: true
        location:
          type: string
        country:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, internship, remote]
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        currency:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        applicantCount:
          type: integer
        employer:
          $ref: "#/components/schemas/User"
          nullable: true
        employerProfile:
          $ref: "#/components/schemas/Profile"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employerId
        - title
        - description
        - location
        - country
        - type
        - skills
        - isActive
        - applicantCount
        - createdAt
        - updatedAt
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - jobs
        - total
        - page
        - totalPages
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        location:
          type: string
        country:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, internship, remote]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
        industry:
          type: string
        skills:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required:
        - title
        - description
        - location
        - country
        - type
        - skills
...
    AiSkillBreakdown:
      type: object
      properties:
        skill:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        yearsExperience:
          type: integer
      required:
        - skill
        - level
        - yearsExperience
    AiReviewResult:
      type: object
      properties:
        overallScore:
          type: integer
          description: Score out of 100
        jobReadinessScore:
          type: integer
        formattingScore:
          type: integer
        skillsScore:
          type: integer
        experienceScore:
          type: integer
        industryAlignmentScore:
          type: integer
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        skillsBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/AiSkillBreakdown"
        careerInsights:
          type: string
        recommendedRoles:
          type: array
          items:
            type: string
      required:
        - overallScore
        - jobReadinessScore
        - formattingScore
        - skillsScore
        - experienceScore
        - industryAlignmentScore
        - summary
        - strengths
        - improvements
        - skillsBreakdown
        - careerInsights
        - recommendedRoles
    HumanReviewResult:
      type: object
      properties:
        professionalFeedback:
          type: string
        structureImprovements:
          type: string
        hiringReadiness:
          type: string
        recommendedRoles:
          type: array
          items:
            type: string
        suggestedImprovements:
          type: array
          items:
            type: string
        reviewerName:
          type: string
        reviewedAt:
          type: string
          format: date-time
      required:
        - professionalFeedback
        - structureImprovements
        - hiringReadiness
        - recommendedRoles
        - suggestedImprovements
        - reviewerName
        - reviewedAt
    CvReview:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        cvFileName:
          type: string
        cvText:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, ai_processing, ai_complete, human_pending, human_processing, human_complete, failed]
        aiReview:
          $ref: "#/components/schemas/AiReviewResult"
          nullable: true
        humanReview:
          $ref: "#/components/schemas/HumanReviewResult"
          nullable: true
        paymentStatus:
          type: string
          enum: [none, pending, paid]
        stripeSessionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - cvFileName
        - status
        - paymentStatus
        - createdAt
        - updatedAt
    CreateCvReviewBody:
      type: object
      properties:
        cvFileName:
          type: string
        cvText:
          type: string
      required:
        - cvFileName
        - cvText
    RequestHumanReviewBody:
      type: object
      properties:
        stripeSessionId:
          type: string
      required:
        - stripeSessionId
...